+++ docs/HISTORY.md
@@ -1,5 +1,13 @@
 # History
 
+## 1.21.0
+
+* Integrate Authentication for Microsoft Accounts (thanks @Kashalls)
+
+## 1.20.2
+
+* add back token auth now that mineflayer is fixed
+
 ## 1.20.1
 
 * Temporarily revert token auth to fix a bug